Revamp /beta to /start
