+++ app/init.js
@@ -66,7 +66,6 @@ module.exports = function(bot) {
     // Enable commands here.
     require('./commands/about.js');
     require('./commands/annokset.js');
-    require('./commands/betablakkis.js');
     require('./commands/help.js');
     require('./commands/jalkikellotus.js');
     require('./commands/juoma.js');
